Launches and runs on Windows

+++ indra/newview/llvisualeffect.cpp
@@ -30,7 +30,7 @@ float LLTweenableValueLerp<float>::get()
 {
     if (!m_CurValue)
     {
-        float curFactor = (LLTimer::getElapsedSeconds() - m_StartTime) / m_Duration;
+        float curFactor = static_cast<float>((LLTimer::getElapsedSeconds() - m_StartTime) / m_Duration);
         if (curFactor < 1.0)
             return lerp(m_StartValue, m_EndValue, curFactor);
         m_CurValue = m_EndValue;
@@ -43,7 +43,7 @@ LLColor3 LLTweenableValueLerp<LLColor3>::get()
 {
     if (!m_CurValue)
     {
-        float curFactor = (LLTimer::getElapsedSeconds() - m_StartTime) / m_Duration;
+        float curFactor = static_cast<float>((LLTimer::getElapsedSeconds() - m_StartTime) / m_Duration);
         if (curFactor < 1.0)
             return lerp(m_StartValue, m_EndValue, curFactor);
         m_CurValue = m_EndValue;
@@ -56,7 +56,7 @@ LLVector4 LLTweenableValueLerp<LLVector4>::get()
 {
     if (!m_CurValue)
     {
-        float curFactor = (LLTimer::getElapsedSeconds() - m_StartTime) / m_Duration;
+        float curFactor = static_cast<float>((LLTimer::getElapsedSeconds() - m_StartTime) / m_Duration);
         if (curFactor < 1.0)
             return lerp(m_StartValue, m_EndValue, curFactor);
         m_CurValue = m_EndValue;
